requirements.txt
Django==1.7c3
Markdown==2.4.1
# Pillow==2.5.3
Pygments==1.6
argparse==1.2.1
certifi==14.05.14
django-filter==0.7
djangorestframework==2.3.14
requests==2.4.0
wsgiref==0.1.2
